refactor(forks): pass fork overrides to clone
