+++ tests/evm_tools/test_fork_cache.py
@@ -13,7 +13,7 @@ from ethereum.fork_criteria import (
     Unscheduled,
 )
 from ethereum_spec_tools.evm_tools.t8n import ForkCache
-from ethereum_spec_tools.forks import Hardfork
+from ethereum_spec_tools.forks import ForkOverrides, Hardfork
 
 pytestmark = pytest.mark.evm_tools
 
@@ -38,6 +38,13 @@ class DummyTemporaryFork:
 def _template() -> Hardfork:
     """Return the Amsterdam fork template."""
     return Hardfork(importlib.import_module("ethereum.forks.amsterdam"))
+
+
+def _seen_overrides(seen: dict[str, Any]) -> ForkOverrides:
+    """Return the ForkOverrides passed to Hardfork.clone."""
+    overrides = seen["overrides"]
+    assert isinstance(overrides, ForkOverrides)
+    return overrides
 
 
 def _different_fork_criteria(
@@ -182,7 +189,7 @@ def test_fork_cache_clones_when_fork_criteria_changes_template(
 
     assert fork is cloned
     assert seen["template"] is template
-    assert seen["fork_criteria"] == changed_fork_criteria
+    assert _seen_overrides(seen).fork_criteria == changed_fork_criteria
 
 
 @pytest.mark.parametrize("field", OVERRIDE_FIELDS)
@@ -228,7 +235,7 @@ def test_fork_cache_clones_for_each_changed_blob_override(
 
     assert fork is cloned
     assert seen["template"] is template
-    assert seen[field] == changed_value
+    assert getattr(_seen_overrides(seen), field) == changed_value
 
 
 def test_fork_cache_reuses_cached_clone_for_identical_changed_request(
